implemented crud for subject and report controllers.

+++ src/main/java/com/dk/repository/SubjectRepository.java
@@ -3,5 +3,5 @@ package com.dk.repository;
 import com.dk.model.Subject;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SubjectRepository extends CrudRepository<Subject, Long> {
+public interface SubjectRepository extends CrudRepository<Subject, Integer> {
 }
